Style Learning Log
Until last commit, tha basic project is fully functional, but it has not styling and can run only on the local computer machine.
In this commit and next few ones, the project will be styled and then deployed to a live server so anyone can make an account and use it.
This commit introduce the django-bootstrap4 app and integrate it into a project to make it ready for live deployment,
including:
1)   The django-bootstrap4 App
2)   Using Bootstrap to Style Learning Log
3)   Modifying base.html
3.a) Defining the HTML Headers
3.b) Defining the Navigation Bar
3.c) Defining the Main Part of the Page
4)   Styling the Home Page Using a Jumbotron
5)   Styling the Login Page
6)   Styling the Topics Page
7)   Styling the Entries on the Topic Page